fix #13
I was very surprised to find that mod filtering is actually unnecessary.
It turns out we can remove it and everything works out magically.

But then I rememberd that filtering was only necessary because of a hack
that we had in place when we didn't have the converter working. Now,
when we have a converter, the hack for a hack can go away.