Reformat everything with scalafmt.
To reformat the project, run ./bin/scalafmt

Why?  Because our time is better spent on solving problems than discuss
trivia. I admit the output from scalafmt in this diff is not always
perfect, but it's good enough for my personal taste.

The CI enforces that new changes are reformatted with scalafmt.