Gets actual class of presentation compiler
Plugin does not switch on in case when IDE actual presentation compiler
inherits from `interactive.Global`. Problem lays in the way how
`analyzer` is replaced with reflection code. Scala for vals creates
private field which is unavailable in descendant class.
So replacing it directly in `interactive.Global` does not override
`analyzer` in actual presentation compiler.